move book list, create etc functionality into separate app called books
