+++ books/urls.py
@@ -1,0 +1,9 @@
+from django.urls import path
+
+from . import views
+
+urlpatterns = [
+    path('', views.index, name='index'),
+    path('books/', views.book_list, name='book_list'),
+    path('book/new', views.book_new, name='book_new'),
+]
