+++ bookx/settings.py
@@ -36,8 +36,9 @@ INSTALLED_APPS = [
     'django.contrib.contenttypes',
     'django.contrib.sessions',
     'django.contrib.messages',
-    'django.contrib.staticfiles',
-    'bookx'
+    'django.contrib.staticfiles',    
+    'users',
+    'books',
 ]
 
 MIDDLEWARE = [
@@ -130,3 +131,5 @@ STATIC_ROOT = os.path.join(BASE_DIR, 'staticfiles')
 
 MEDIA_URL = '/media/'
 MEDIA_ROOT = os.path.join(BASE_DIR, 'media')
+
+AUT_USER_MODEL = 'users.CustomUser'
