+++ templates/books/book_list.html
@@ -1,4 +1,4 @@
-{% extends 'bookx/base.html' %}
+{% extends 'books/base.html' %}
 
 {% block content %}
