+++ templates/books/signup.html
@@ -1,4 +1,4 @@
-{% extends 'base.html' %}
+{% extends 'books/base.html' %}
 
 {% block content %}
   <h2>Sign up</h2>
